Third commit
The incog_tangle, prob_statistic, paraF and gini_plot functions are added, along with their descriptions and arguments, but the examples are still missing. I update the descriptions and examples of the rest of the functions. I add two databases: birds_mites and plant_fungi (although the second one will surely have to be replaced). For now, the package has no errors messages, but it does have 4 notes and 3 warnings, most of them coming from the lack of the examples in the new functions (upload in the next commit).