Deep package modification and renaming
I have removed some of the functions and, for now, also removed the databases until some checks are done. At the moment, the examples cannot be run. The incong_tangle function remains, but will probably be removed.
Main functions like timsHS_max(C or I), link_freq and those corresponding to the global-fit methods have been modified. link_freq has the biggest change, now allowing analysis of frequencies at both the 0.01 and 0.99 percentiles without errors.
Two functions have been added (max_cong and max_incong) which allow to apply the maximum congruence algorithm or the maximum incongruence algorithm with a single execution, including the global-fit methods described in the package. In addition, an argument has been added, diff.fq, which allows calculating the differences in the frequencies between the 0.99 and 0.01 percentiles, detecting the most "remarkable" links in the phylogenies. In addition, it partially corrects the overestimation and tendency to choose one-to-one links by maximizing congruence, and the underestimation and tendency to choose multiple links by maximizing incongruence.
It remains to modify the prob_statistic function (include a note about its computational load, more in certain datasets than others, and also depending on the global-fit method used). max_cong and max_incong have to be modified to accept any global-fit method when diff.fq = TRUE (only PACo is accepted at the moment).