4th commit
I have modified the parallelization part of the paraF function (to decrease the computational time). I have also modified the prob_statistic function to reduce the execution time as much as possible.
In the NAMESPACE, I have modified the imported packages, to minimize its load. I have added the project folder (.Rproj) to .gitignore because it was causing problems with commits.
The datasets and examples have not yet been updated. It is the next step.